[model-gateway] refactor oai streaming -2
