+++ sgl-model-gateway/src/routers/openai/mod.rs
@@ -7,6 +7,7 @@
 //! - Multi-turn tool execution loops
 //! - SSE (Server-Sent Events) streaming
 
+mod accumulator;
 mod context;
 pub mod conversations;
 pub mod mcp;
@@ -14,6 +15,7 @@ pub mod provider;
 mod responses;
 mod router;
 mod streaming;
+mod tool_handler;
 
 // Re-export the main types for external use
 pub use provider::{Provider, ProviderError, ProviderRegistry};
